+++ NEWS.md
@@ -1,7 +1,8 @@
 # formods 0.1.2
 
 * development version
-* Added `new_module_template()` to create new module templates. 
+* Added `new_module_template()` to create new module templates and `use_formods()` 
+to automatically add the files to a package. 
 
 # formods 0.1.1